+++ PayPalMobile/PayPalConfiguration.h
@@ -1,0 +1,92 @@
+//
+//  PayPalConfiguration.h
+//
+//  Version 2.0.0
+//
+//  Copyright (c) 2013, PayPal
+//  All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+/// You use a PayPalConfiguration object to configure many aspects of how the SDK behaves.
+/// The PayPalConfiguration class also includes two important class methods.
+@interface PayPalConfiguration : NSObject <NSCopying>
+
+/// Optional default user email address to be shown on the PayPal login view.
+/// Will be overridden by email used in most recent PayPal login.
+@property(nonatomic, copy, readwrite) NSString *defaultUserEmail;
+/// Optional default user phone country code used in the PayPal login view.
+/// Will be overridden by phone country code used in most recent PayPal login.
+@property(nonatomic, copy, readwrite) NSString *defaultUserPhoneCountryCode;
+/// Optional default user phone number to be shown in the PayPal login view.
+/// Will be overridden by phone number used in most recent PayPal login.
+/// @note If you set defaultUserPhoneNumber, be sure to also set defaultUserPhoneCountryCode.
+@property(nonatomic, copy, readwrite) NSString *defaultUserPhoneNumber;
+
+/// Your company name, as it should be displayed to the user
+/// when requesting consent via a PayPalFuturePaymentViewController.
+@property(nonatomic, copy, readwrite) NSString *merchantName;
+/// URL of your company's privacy policy, which will be offered to the user
+/// when requesting consent via a PayPalFuturePaymentViewController.
+@property(nonatomic, copy, readwrite) NSURL *merchantPrivacyPolicyURL;
+/// URL of your company's user agreement, which will be offered to the user
+/// when requesting consent via a PayPalFuturePaymentViewController.
+@property(nonatomic, copy, readwrite) NSURL *merchantUserAgreementURL;
+
+/// If set to NO, the SDK will only support paying with PayPal, not with credit cards.
+/// This applies only to single payments (via PayPalPaymentViewController).
+/// Future payments (via PayPalFuturePaymentViewController) always use PayPal.
+/// Defaults to YES.
+@property(nonatomic, assign, readwrite) BOOL acceptCreditCards;
+
+/// If set to YES, then if the user pays via their PayPal account,
+/// the SDK will remember the user's PayPal username or phone number;
+/// if the user pays via their credit card, then the SDK will remember
+/// the PayPal Vault token representing the user's credit card.
+///
+/// If set to NO, then any previously-remembered username, phone number, or
+/// credit card token will be erased, and subsequent payment information will
+/// not be remembered.
+///
+/// Defaults to YES.
+@property(nonatomic, assign, readwrite) BOOL rememberUser;
+
+/// If not set, or if set to nil, defaults to the device's current language setting.
+///
+/// Can be specified as a language code ("en", "fr", "zh-Hans", etc.) or as a locale ("en_AU", "fr_FR", "zh-Hant_HK", etc.).
+/// If the library does not contain localized strings for a specified locale, then will fall back to the language. E.g., "es_CO" -> "es".
+/// If the library does not contain localized strings for a specified language, then will fall back to American English.
+///
+/// If you specify only a language code, and that code matches the device's currently preferred language,
+/// then the library will attempt to use the device's current region as well.
+/// E.g., specifying "en" on a device set to "English" and "United Kingdom" will result in "en_GB".
+///
+/// These localizations are currently included:
+/// da,de,en,en_AU,en_GB,en_SV,es,es_MX,fr,he,it,ja,ko,nb,nl,pl,pt,pt_BR,ru,sv,tr,zh-Hans,zh-Hant_HK,zh-Hant_TW.
+@property(nonatomic, copy, readwrite) NSString *languageOrLocale;
+#if DEBUG
+
+/// For debug experimentation only!
+/// If this property is nil, then PPServiceManager will fall back to usual logic for determining adaptation country.
+@property (nonatomic, copy, readwrite) NSString *countryForAdaptation;
+#endif
+
+/// Normally, the SDK blurs the screen when the app is backgrounded,
+/// to obscure credit card or PayPal account details in the iOS-saved screenshot.
+/// If your app already does its own blurring upon backgrounding, you might choose to disable this.
+/// Defaults to NO.
+@property(nonatomic, assign, readwrite) BOOL disableBlurWhenBackgrounding;
+
+/// Sandbox credentials can be difficult to type on a mobile device. Setting this flag to YES will
+/// cause the sandboxUserPassword and sandboxUserPin to always be pre-populated into login fields.
+///
+/// This setting will have no effect if the operation mode is production.
+@property (nonatomic, assign, readwrite) BOOL forceDefaultsInSandbox;
+
+/// Password to use for sandbox if 'forceDefaultsInSandbox' is set.
+@property (nonatomic, copy, readwrite) NSString* sandboxUserPassword;
+/// PIN to use for sandbox if 'forceDefaultsInSandbox' is set.
+@property (nonatomic, copy, readwrite) NSString* sandboxUserPin;
+
+@end