+++ PayPalMobile/PayPalPayment.h
@@ -1,37 +1,34 @@
 //
 //  PayPalPayment.h
 //
-//  Version 1.4.6
+//  Version 2.0.0
 //
 //  Copyright (c) 2013, PayPal
 //  All rights reserved.
 //
-//  Redistribution and use in source and binary forms, with or without
-//  modification, are permitted provided that the following conditions are met:
-//
-//  1. Redistributions of source code must retain the above copyright notice, this
-//  list of conditions and the following disclaimer.
-//  2. Redistributions in binary form must reproduce the above copyright notice,
-//  this list of conditions and the following disclaimer in the documentation
-//  and/or other materials provided with the distribution.
-//
-//  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-//  ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-//  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-//  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
-//  ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-//  (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-//   LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-//  ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-//  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-//  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-//  The views and conclusions contained in the software and documentation are those
-//  of the authors and should not be interpreted as representing official policies,
-//  either expressed or implied, of the FreeBSD Project.
-//
 
 #import <Foundation/Foundation.h>
+
+typedef NS_ENUM(NSInteger, PayPalPaymentIntent) {
+  PayPalPaymentIntentSale = 0,
+  PayPalPaymentIntentAuthorize = 1,
+  };
+
+/// The PayPalPaymentDetails class defines optional amount details.
+/// @see https://developer.paypal.com/webapps/developer/docs/api/#details-object for more details.
+@interface PayPalPaymentDetails : NSObject <NSCopying>
+
+/// Amount charged for shipping. 10 characters max with support for 2 decimal places.
+@property(nonatomic, copy, readwrite) NSDecimalNumber *shipping;
+
+/// Sub-total (amount) of items being paid for. 10 characters max with support for 2 decimal places.
+@property(nonatomic, copy, readwrite) NSDecimalNumber *subtotal;
+
+/// Amount charged for tax. 10 characters max with support for 2 decimal places.
+@property(nonatomic, copy, readwrite) NSDecimalNumber *tax;
+
+@end
+
 
 @interface PayPalPayment : NSObject <NSCopying>
 
@@ -40,9 +37,12 @@
 /// @param amount The amount of the payment.
 /// @param currencyCode The ISO 4217 currency for the payment.
 /// @param shortDescription A short descripton of the payment.
+/// @param intent PayPalPaymentIntentSale for an immediate payment;
+///                PayPalPaymentIntentAuthorize for payment authorization only, to be captured separately at a later time.
 + (PayPalPayment *)paymentWithAmount:(NSDecimalNumber *)amount
                         currencyCode:(NSString *)currencyCode
-                    shortDescription:(NSString *)shortDescription;
+                    shortDescription:(NSString *)shortDescription
+                              intent:(PayPalPaymentIntent)intent;
 
 
 #pragma mark - Required properties
@@ -57,6 +57,23 @@
 /// The description will be truncated for display, if necessary;
 /// limiting it to about 20 characters should prevent truncation in most cases.
 @property(nonatomic, copy, readwrite) NSString *shortDescription;
+
+/// The intent of this payment:
+/// - PayPalPaymentIntentSale for an immediate payment
+/// - PayPalPaymentIntentAuthorize for payment authorization only,
+///   to be captured separately at a later time.
+/// Defaults to PayPalPaymentIntentSale.
+@property(nonatomic, assign, readwrite) PayPalPaymentIntent intent;
+
+
+#pragma mark - Optional properties
+
+/// Optional payment details @see PaymentDetails properties.
+@property (nonatomic, copy, readwrite) PayPalPaymentDetails *paymentDetails;
+
+/// Optional Build Notation code ("BN code"), obtained from partnerprogram@paypal.com,
+/// for your tracking purposes.
+@property(nonatomic, copy, readwrite) NSString *bnCode;
 
 
 #pragma mark - Convenience getters