+++ SampleApp/PayPal-iOS-SDK-Sample-App/ZZFlipsideViewController.m
@@ -5,40 +5,17 @@
 //  Copyright (c) 2013, PayPal
 //  All rights reserved.
 //
-//  Redistribution and use in source and binary forms, with or without
-//  modification, are permitted provided that the following conditions are met:
-//
-//  1. Redistributions of source code must retain the above copyright notice, this
-//  list of conditions and the following disclaimer.
-//  2. Redistributions in binary form must reproduce the above copyright notice,
-//  this list of conditions and the following disclaimer in the documentation
-//  and/or other materials provided with the distribution.
-//
-//  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-//  ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-//  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-//  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
-//  ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-//  (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-//   LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-//  ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-//  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-//  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-//  The views and conclusions contained in the software and documentation are those
-//  of the authors and should not be interpreted as representing official policies,
-//  either expressed or implied, of the FreeBSD Project.
-//
 
 #import "ZZFlipsideViewController.h"
+#import "ZZMainViewController.h"
 #import <QuartzCore/QuartzCore.h>
 
 @interface ZZFlipsideViewController ()
 
 @property(nonatomic, strong, readwrite) IBOutlet UISegmentedControl *environmentSegmentedControl;
 @property(nonatomic, strong, readwrite) IBOutlet UISwitch *acceptCreditCardsSwitch;
-@property(nonatomic, strong, readwrite) IBOutlet UITextView *proofOfPaymentTextView;
-@property(nonatomic, strong, readwrite) IBOutlet UILabel *proofOfPaymentLabel;
+@property(nonatomic, strong, readwrite) IBOutlet UITextView *payPalResultTextView;
+@property(nonatomic, strong, readwrite) IBOutlet UILabel *payPalResultLabel;
 @end
 
 @implementation ZZFlipsideViewController
@@ -49,6 +26,7 @@
 }
 
 - (void)viewDidLoad {
+
   [super viewDidLoad];
   
 }
@@ -59,25 +37,25 @@
 
 - (void)viewWillAppear:(BOOL)animated {
   [self logEnvironment];
-  if ([PayPalEnvironmentProduction isEqualToString:self.delegate.environment]) {
-    self.environmentSegmentedControl.selectedSegmentIndex = 2;
-  } else if ([PayPalEnvironmentSandbox isEqualToString:self.delegate.environment]) {
-    self.environmentSegmentedControl.selectedSegmentIndex = 1;
-  } else if ([PayPalEnvironmentNoNetwork isEqualToString:self.delegate.environment]) {
-    self.environmentSegmentedControl.selectedSegmentIndex = 0;
+
+  int numberOfSegments = self.environmentSegmentedControl.numberOfSegments;
+  while (numberOfSegments--) {
+    NSString *title = [self.environmentSegmentedControl titleForSegmentAtIndex:numberOfSegments];
+    if ([[title lowercaseString] isEqualToString:self.delegate.environment]) {
+      self.environmentSegmentedControl.selectedSegmentIndex = numberOfSegments;
+      break;
+    }
   }
   self.acceptCreditCardsSwitch.on = self.delegate.acceptCreditCards;
 
-  if ([self.delegate completedPayment]) {
-    NSLog(@"%@", [self.delegate completedPayment]);
-    self.proofOfPaymentTextView.text = [[NSString stringWithFormat:@"%@", [self.delegate completedPayment]] stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
-    
-    
+  if ([self.delegate resultText]) {
+    NSLog(@"%@", [self.delegate resultText]);
+    self.payPalResultTextView.text = [[self.delegate resultText] stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
   } else {
-    self.proofOfPaymentTextView.hidden = YES;
-    self.proofOfPaymentLabel.hidden = YES;
+    self.payPalResultTextView.hidden = YES;
+    self.payPalResultLabel.hidden = YES;
   }
-  self.proofOfPaymentTextView.layer.cornerRadius = 8.0f;
+  self.payPalResultTextView.layer.cornerRadius = 8.0f;
 }
 
 - (void)didReceiveMemoryWarning {
@@ -88,16 +66,8 @@
 #pragma mark - Actions
 
 - (IBAction)environmentControlDidUpdate:(id)sender {
-  switch (self.environmentSegmentedControl.selectedSegmentIndex) {
-    case 0:
-      self.delegate.environment = PayPalEnvironmentNoNetwork;
-      break;
-    case 1:
-      self.delegate.environment = PayPalEnvironmentSandbox;
-      break;
-    default:
-      self.delegate.environment = PayPalEnvironmentProduction;
-      break;
+  if (self.environmentSegmentedControl == sender) {
+    self.delegate.environment = [[self.environmentSegmentedControl titleForSegmentAtIndex:self.environmentSegmentedControl.selectedSegmentIndex] lowercaseString];
   }
   [self logEnvironment];
 }
@@ -110,10 +80,6 @@
 - (IBAction)done:(id)sender {
   [self logEnvironment];
   [self.delegate flipsideViewControllerDidFinish:self];
-}
-
-- (BOOL)prefersStatusBarHidden {
-  return YES;
 }
 
 @end