+++ SampleApp/PayPal-iOS-SDK-Sample-App/ZZMainViewController.m
@@ -5,42 +5,23 @@
 //  Copyright (c) 2013, PayPal
 //  All rights reserved.
 //
-//  Redistribution and use in source and binary forms, with or without
-//  modification, are permitted provided that the following conditions are met:
-//
-//  1. Redistributions of source code must retain the above copyright notice, this
-//  list of conditions and the following disclaimer.
-//  2. Redistributions in binary form must reproduce the above copyright notice,
-//  this list of conditions and the following disclaimer in the documentation
-//  and/or other materials provided with the distribution.
-//
-//  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-//  ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-//  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-//  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
-//  ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-//  (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-//   LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-//  ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-//  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-//  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-//  The views and conclusions contained in the software and documentation are those
-//  of the authors and should not be interpreted as representing official policies,
-//  either expressed or implied, of the FreeBSD Project.
-//
 
 #import "ZZMainViewController.h"
 #import <QuartzCore/QuartzCore.h>
 
-#warning "Enter your credentials"
-#define kPayPalClientId @"YOUR CLIENT ID HERE"
-#define kPayPalReceiverEmail @"YOUR_PAYPAL_EMAIL@yourdomain.com"
+// Set the environment:
+// - For live charges, use PayPalEnvironmentProduction (default).
+// - To use the PayPal sandbox, use PayPalEnvironmentSandbox.
+// - For testing, use PayPalEnvironmentNoNetwork.
+#define kPayPalEnvironment PayPalEnvironmentNoNetwork
 
 @interface ZZMainViewController ()
 
-@property(nonatomic, strong, readwrite) IBOutlet UIButton *payButton;
+@property(nonatomic, strong, readwrite) IBOutlet UIButton *payNowButton;
+@property(nonatomic, strong, readwrite) IBOutlet UIButton *payFutureButton;
 @property(nonatomic, strong, readwrite) IBOutlet UIView *successView;
+
+@property(nonatomic, strong, readwrite) PayPalConfiguration *payPalConfig;
 
 @end
 
@@ -48,43 +29,52 @@
 
 - (void)viewDidLoad {
   [super viewDidLoad];
-  self.title = @"PayPal iOS Library Demo";
-  self.acceptCreditCards = YES;
-  self.environment = PayPalEnvironmentNoNetwork;
+  self.title = @"PayPal SDK Demo";
+
+  // Set up payPalConfig
+  _payPalConfig = [[PayPalConfiguration alloc] init];
+  _payPalConfig.acceptCreditCards = YES;
+  _payPalConfig.languageOrLocale = @"en";
+  _payPalConfig.merchantName = @"Awesome Shirts, Inc.";
+  _payPalConfig.merchantPrivacyPolicyURL = [NSURL URLWithString:@"https://www.paypal.com/webapps/mpp/ua/privacy-full"];
+  _payPalConfig.merchantUserAgreementURL = [NSURL URLWithString:@"https://www.paypal.com/webapps/mpp/ua/useragreement-full"];
+  
+  // Setting the languageOrLocale property is optional.
+  //
+  // If you do not set languageOrLocale, then the PayPalPaymentViewController will present
+  // its user interface according to the device's current language setting.
+  //
+  // Setting languageOrLocale to a particular language (e.g., @"es" for Spanish) or
+  // locale (e.g., @"es_MX" for Mexican Spanish) forces the PayPalPaymentViewController
+  // to use that language/locale.
+  //
+  // For full details, including a list of available languages and locales, see PayPalPaymentViewController.h.
+  
+  _payPalConfig.languageOrLocale = [NSLocale preferredLanguages][0];
+
   // Do any additional setup after loading the view, typically from a nib.
 
   self.successView.hidden = YES;
+  
+  // use default environment, should be Production in real life
+  self.environment = kPayPalEnvironment;
 
-  NSLog(@"PayPal iOS SDK version: %@", [PayPalPaymentViewController libraryVersion]);
+  NSLog(@"PayPal iOS SDK version: %@", [PayPalMobile libraryVersion]);
+
 }
 
 - (void)viewWillAppear:(BOOL)animated {
   [super viewWillAppear:YES];
 
-  UIEdgeInsets insets = UIEdgeInsetsMake(0, 15.0f, 0, 14.0f);
-  UIImage *payBackgroundImage = [[UIImage imageNamed:@"button_secondary.png"] resizableImageWithCapInsets:insets];
-  UIImage *payBackgroundImageHighlighted = [[UIImage imageNamed:@"button_secondary_selected.png"] resizableImageWithCapInsets:insets];
-  [self.payButton setBackgroundImage:payBackgroundImage forState:UIControlStateNormal];
-  [self.payButton setBackgroundImage:payBackgroundImageHighlighted forState:UIControlStateHighlighted];
-  [self.payButton setTitleColor:[UIColor darkGrayColor] forState:UIControlStateNormal];
-  [self.payButton setTitleColor:[UIColor darkGrayColor] forState:UIControlStateHighlighted];
-  
-  // Optimization: Prepare for display of the payment UI by getting network work done early
-  [PayPalPaymentViewController setEnvironment:self.environment];
-  [PayPalPaymentViewController prepareForPaymentUsingClientId:kPayPalClientId];
+  // Preconnect to PayPal early
+  [PayPalMobile preconnectWithEnvironment:self.environment];
 }
 
-- (void)didReceiveMemoryWarning {
-  [super didReceiveMemoryWarning];
-  // Dispose of any resources that can be recreated.
-}
-
-#pragma mark - Pay action
+#pragma mark - Receive Single Payment
 
 - (IBAction)pay {
-
   // Remove our last completed payment, just for demo purposes.
-  self.completedPayment = nil;
+  self.resultText = nil;
   
   PayPalPayment *payment = [[PayPalPayment alloc] init];
   payment.amount = [[NSDecimalNumber alloc] initWithString:@"9.95"];
@@ -98,61 +88,83 @@
     // to handle that here.
   }
 
-  // Any customer identifier that you have will work here. Do NOT use a device- or
-  // hardware-based identifier.
-  NSString *customerId = @"user-11723";
-
-  // Set the environment:
-  // - For live charges, use PayPalEnvironmentProduction (default).
-  // - To use the PayPal sandbox, use PayPalEnvironmentSandbox.
-  // - For testing, use PayPalEnvironmentNoNetwork.
-  [PayPalPaymentViewController setEnvironment:self.environment];
-
-  PayPalPaymentViewController *paymentViewController = [[PayPalPaymentViewController alloc] initWithClientId:kPayPalClientId
-                                                                                               receiverEmail:kPayPalReceiverEmail
-                                                                                                     payerId:customerId
-                                                                                                     payment:payment
-                                                                                                    delegate:self];
-  paymentViewController.hideCreditCardButton = !self.acceptCreditCards;
+  // Update payPalConfig re accepting credit cards.
+  self.payPalConfig.acceptCreditCards = self.acceptCreditCards;
   
-  // Setting the languageOrLocale property is optional.
-  //
-  // If you do not set languageOrLocale, then the PayPalPaymentViewController will present
-  // its user interface according to the device's current language setting.
-  //
-  // Setting languageOrLocale to a particular language (e.g., @"es" for Spanish) or
-  // locale (e.g., @"es_MX" for Mexican Spanish) forces the PayPalPaymentViewController
-  // to use that language/locale.
-  //
-  // For full details, including a list of available languages and locales, see PayPalPaymentViewController.h.
-  paymentViewController.languageOrLocale = @"en";
-
+  PayPalPaymentViewController *paymentViewController = [[PayPalPaymentViewController alloc] initWithPayment:payment
+                                                                                              configuration:self.payPalConfig
+                                                                                                   delegate:self];
   [self presentViewController:paymentViewController animated:YES completion:nil];
 }
 
-#pragma mark - Proof of payment validation
+#pragma mark PayPalPaymentDelegate methods
+
+- (void)payPalPaymentViewController:(PayPalPaymentViewController *)paymentViewController didCompletePayment:(PayPalPayment *)completedPayment {
+  NSLog(@"PayPal Payment Success!");
+  self.resultText = [completedPayment description];
+  [self showSuccess];
+
+  [self sendCompletedPaymentToServer:completedPayment]; // Payment was processed successfully; send to server for verification and fulfillment
+  [self dismissViewControllerAnimated:YES completion:nil];
+}
+
+- (void)payPalPaymentDidCancel:(PayPalPaymentViewController *)paymentViewController {
+  NSLog(@"PayPal Payment Canceled");
+  self.resultText = nil;
+  self.successView.hidden = YES;
+  [self dismissViewControllerAnimated:YES completion:nil];
+}
+
+#pragma mark Proof of payment validation
 
 - (void)sendCompletedPaymentToServer:(PayPalPayment *)completedPayment {
   // TODO: Send completedPayment.confirmation to server
   NSLog(@"Here is your proof of payment:\n\n%@\n\nSend this to your server for confirmation and fulfillment.", completedPayment.confirmation);
 }
 
-#pragma mark - PayPalPaymentDelegate methods
 
-- (void)payPalPaymentDidComplete:(PayPalPayment *)completedPayment {
-  NSLog(@"PayPal Payment Success!");
-  self.completedPayment = completedPayment;
-  self.successView.hidden = NO;
+#pragma mark - Authorize Future Payments
+
+- (IBAction)getUserAuthorization:(id)sender {
   
-  [self sendCompletedPaymentToServer:completedPayment]; // Payment was processed successfully; send to server for verification and fulfillment
+  PayPalFuturePaymentViewController *futurePaymentViewController = [[PayPalFuturePaymentViewController alloc] initWithConfiguration:self.payPalConfig delegate:self];
+  [self presentViewController:futurePaymentViewController animated:YES completion:nil];
+}
+
+
+#pragma mark PayPalFuturePaymentDelegate methods
+
+- (void)payPalFuturePaymentViewController:(PayPalFuturePaymentViewController *)futurePaymentViewController didAuthorizeFuturePayment:(NSDictionary *)futurePaymentAuthorization {
+  NSLog(@"PayPal Future Payment Authorization Success!");
+  self.resultText = futurePaymentAuthorization[@"code"];
+  [self showSuccess];
+
+  [self sendAuthorizationToServer:futurePaymentAuthorization];
   [self dismissViewControllerAnimated:YES completion:nil];
 }
 
-- (void)payPalPaymentDidCancel {
-  NSLog(@"PayPal Payment Canceled");
-  self.completedPayment = nil;
+- (void)payPalFuturePaymentDidCancel:(PayPalFuturePaymentViewController *)futurePaymentViewController {
+  NSLog(@"PayPal Future Payment Authorization Canceled");
   self.successView.hidden = YES;
   [self dismissViewControllerAnimated:YES completion:nil];
+}
+
+- (void)sendAuthorizationToServer:(NSDictionary *)authorization {
+  // TODO: Send authorization to server
+  NSLog(@"Here is your authorization:\n\n%@\n\nSend this to your server to complete future payment setup.", authorization);
+}
+
+
+#pragma mark - Helpers
+
+- (void)showSuccess {
+  self.successView.hidden = NO;
+  self.successView.alpha = 1.0f;
+  [UIView beginAnimations:nil context:NULL];
+  [UIView setAnimationDuration:0.5];
+  [UIView setAnimationDelay:2.0];
+  self.successView.alpha = 0.0f;
+  [UIView commitAnimations];
 }
 
 #pragma mark - Flipside View Controller
@@ -171,7 +183,7 @@
 }
 
 - (void)prepareForSegue:(UIStoryboardSegue *)segue sender:(id)sender {
-  if ([[segue identifier] isEqualToString:@"showAlternate"]) {
+  if ([[segue identifier] isEqualToString:@"pushSettings"]) {
     [[segue destinationViewController] setDelegate:self];
 
     if ([[UIDevice currentDevice] userInterfaceIdiom] == UIUserInterfaceIdiomPad) {
@@ -189,10 +201,6 @@
   } else {
     [self performSegueWithIdentifier:@"showAlternate" sender:sender];
   }
-}
-
-- (BOOL)prefersStatusBarHidden {
-  return YES;
 }
 
 @end